refactor(shared): Fase 1b — 8 globais universais -> rts-shared
Batch limpo de globais que so tocam engine heap (+ boolean: string_pool extern):
boolean bigint number url weakmap weakset weakref finalization_registry.

- git mv globals/{8} -> rts-shared/src/globals/
- gc::handles -> rts_engine::heap::handles (7; bigint nao usa heap)
- boolean: gc::string_pool::__RTS_FN_NS_GC_STRING_NEW -> extern decl (unsafe)
- rts-shared globals/mod.rs: pub mod p/ os 8
- facade: namespaces/globals/mod.rs pub use rts_shared::globals::<g>
- codegen register (g::<g>::register_*_class_spec) resolve via facade, sem mudanca

Restam no runtime (coupling pub-fn helper/cluster): reflect(collections::map),
e o cluster function/collections/proxy/error/string (precisa relocar pure
helpers format_js_number + stack_for_handle do collector antes).

Gate: cargo build --release OK (sem warnings); rts.exe test 1710/1710.
