+++ crates/rts-shared/src/globals/bigint/mod.rs
