+++ crates/rts-shared/src/globals/weakmap/mod.rs
@@ -13,7 +13,7 @@ use std::collections::HashMap;
 use rts_engine::abi::ty::{Bool, Handle, I64};
 use rts_engine::{AbiType, Engine, FnPtr, Member, MemberFlags, MemberKind, Sig};
 
-use crate::namespaces::gc::handles::{alloc_entry, with_entry, with_entry_mut, Entry};
+use rts_engine::heap::handles::{alloc_entry, with_entry, with_entry_mut, Entry};
 
 /// Creates a new empty WeakMap.
 #[unsafe(no_mangle)]
