+++ crates/rts-shared/src/globals/weakref/mod.rs
@@ -9,7 +9,7 @@
 use rts_engine::abi::ty::Handle;
 use rts_engine::{AbiType, Engine, FnPtr, Member, MemberFlags, MemberKind, Sig};
 
-use crate::namespaces::gc::handles::{alloc_entry, with_entry, Entry};
+use rts_engine::heap::handles::{alloc_entry, with_entry, Entry};
 
 /// Creates a new WeakRef wrapping target.
 #[unsafe(no_mangle)]
